split CustomSoundsManager off of ProfileManager
