+++ app/src/main/java/ie/delilahsthings/soothingloop/Util.java
@@ -1,9 +1,26 @@
 package ie.delilahsthings.soothingloop;
 
+import android.annotation.SuppressLint;
+import android.content.Context;
+import android.database.Cursor;
+import android.net.Uri;
+import android.provider.OpenableColumns;
 import android.view.View;
 import android.view.ViewGroup;
 
+import java.io.File;
+
 public abstract class Util {
+    public static String[] safe_list(File dir) {
+        String[] lst = dir.list();
+        if(lst==null){
+            return new String[0];
+        }
+        else {
+            return lst;
+        }
+    }
+
     static void sleep(long milis)
     {
         try {
@@ -26,5 +43,28 @@ public abstract class Util {
         }
 
         return null;
+    }
+
+    @SuppressLint("Range")
+    public static String getFileName(Context context, Uri uri) {
+        String result = null;
+        if (uri.getScheme().equals("content")) {
+            Cursor cursor = context.getContentResolver().query(uri, null, null, null, null);
+            try {
+                if (cursor != null && cursor.moveToFirst()) {
+                    result = cursor.getString(cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME));
+                }
+            } finally {
+                cursor.close();
+            }
+        }
+        if (result == null) {
+            result = uri.getPath();
+            int cut = result.lastIndexOf('/');
+            if (cut != -1) {
+                result = result.substring(cut + 1);
+            }
+        }
+        return result;
     }
 }
